requirements.txt
aiofiles==0.8.0
anyio==3.5.0
argcomplete==2.0.0
autocfg==0.0.8
backports.zoneinfo==0.2.1
boto3==1.21.3
botocore==1.24.3
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.0
eventlet==0.33.0
fiftyone==0.14.4
fiftyone-brain==0.8.0
fiftyone-db==0.3.0
fonttools==4.29.1
future==0.18.2
glob2==0.7
gluoncv==0.10.4.post4
graphviz==0.8.4
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
imageio==2.16.0
importlib-metadata==4.11.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
mongoengine==0.20.0
motor==2.5.1
mxnet==1.8.0.post0
ndjson==0.3.1
networkx==2.6.3
numpy==1.21.5
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.3.5
patool==1.12
Pillow==9.0.1
plotly==4.14.3
portalocker==2.4.0
pprintpp==0.4.0
psutil==5.9.0
pycocotools==2.0.4
pymongo==3.12.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
requests==2.27.1
retrying==1.3.3
rfc3986==1.5.0
s3transfer==0.5.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tabulate==0.8.9
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
universal-analytics-python3==1.1.1
urllib3==1.26.8
utils==1.0.1
voxel51-eta==0.6.4
wrapt==1.13.3
xmltodict==0.12.0
yacs==0.1.8
zipp==3.7.0
